+++ vsfieldkit/__init__.py
@@ -7,7 +7,7 @@ from vsfieldkit.types import (ChromaSubsampleScanning, Factor, FormatSpecifier,
 from vsfieldkit.util import (assume_bff, assume_progressive, assume_tff,
                              double, group_by_combed, group_by_field_order)
 
-VERSION = 1, 0, 1
+VERSION = 1, 0, 2
 
 SCAN_BLENDED = ChromaSubsampleScanning.SCAN_BLENDED
 SCAN_LATEST = ChromaSubsampleScanning.SCAN_LATEST