+++ docs/changelog.rst
@@ -1,5 +1,11 @@
 Changelog
 =========
+1.0.2
+------------
+* :py:func:`vsfieldkit.fill_analog_frame_ends` will now look for EdgeFixer
+  plugin first, followed by ContinuityFixer plugin as before. Having one of the
+  two plugins is required.
+
 1.0.1
 ------------
 * Adds :py:func:`vsfieldkit.fill_analog_frame_ends` for cleaning the half-line
@@ -8,8 +14,8 @@ Changelog
 Output Change:
 
 * :py:func:`vsfieldkit.bob` now defaults to shifting according to the field's
-  position. Feature added for completion, but it's also deprecated
-  in favor of :py:func:`resize.Bob` in VapourSynth R58+.
+  position. Feature added for completion, but it's also deprecated in favor of
+  :py:func:`resize.Bob` in VapourSynth R58+.
 
 Version 1.0.0 was yanked for an immediate bug fix.